+++ src/actions/submit/submit_action.ts
@@ -4,7 +4,7 @@ import { TContext } from '../../lib/context';
 import { TScopeSpec } from '../../lib/engine/scope_spec';
 import { ExitFailedError, KilledError } from '../../lib/errors';
 import { cliAuthPrecondition } from '../../lib/preconditions';
-import { getSurvey, showSurvey } from '../../lib/telemetry/survey/survey';
+import { getSurvey, showSurvey } from '../survey';
 import { getPRInfoForBranches } from './prepare_branches';
 import { submitPullRequest } from './submit_prs';
 import { validateBranchesToSubmit } from './validate_branches';
